iprt/getopt.h/cpp: Made the uIndex 32-bit instead of 64-bit as we don't need 64-bit and it only causes lots of warnings when compiling VBoxManage with MSC.
svn:sync-xref-src-repo-rev: r55902

+++ src/VBox/Runtime/testcase/tstGetOpt.cpp
@@ -1,4 +1,4 @@
-/* $Id: tstGetOpt.cpp 24140 2009-10-28 14:29:34Z noreply@oracle.com $ */
+/* $Id: tstGetOpt.cpp 25323 2009-12-11 12:27:17Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT Testcase - RTGetOpt
  */
@@ -171,10 +171,10 @@ int main()
 
     CHECK_GETOPT(RTGetOpt(&GetState, &Val), 's', 2);
     CHECK(VALID_PTR(Val.psz) && !strcmp(Val.psz, "string1"));
-    CHECK(GetState.uIndex == UINT64_MAX);
+    CHECK(GetState.uIndex == UINT32_MAX);
     CHECK_GETOPT(RTGetOpt(&GetState, &Val), 's', 2);
     CHECK(VALID_PTR(Val.psz) && !strcmp(Val.psz, "string2"));
-    CHECK(GetState.uIndex == UINT64_MAX);
+    CHECK(GetState.uIndex == UINT32_MAX);
 
     /* -i */
     CHECK_GETOPT(RTGetOpt(&GetState, &Val), 'i', 2);
@@ -300,22 +300,22 @@ int main()
     CHECK(GetState.uIndex == 1);
 
     CHECK_GETOPT(RTGetOpt(&GetState, &Val), 404, 1);
-    CHECK(GetState.uIndex == UINT64_MAX);
+    CHECK(GetState.uIndex == UINT32_MAX);
 
     /* RTGetOptFetchValue tests */
     CHECK_GETOPT(RTGetOpt(&GetState, &Val), 405, 2);
     CHECK(VALID_PTR(Val.psz) && !strcmp(Val.psz, "firstvalue"));
-    CHECK(GetState.uIndex == UINT64_MAX);
+    CHECK(GetState.uIndex == UINT32_MAX);
     CHECK_GETOPT(RTGetOptFetchValue(&GetState, &Val, RTGETOPT_REQ_STRING), VINF_SUCCESS, 1);
     CHECK(VALID_PTR(Val.psz) && !strcmp(Val.psz, "secondvalue"));
-    CHECK(GetState.uIndex == UINT64_MAX);
+    CHECK(GetState.uIndex == UINT32_MAX);
 
     CHECK_GETOPT(RTGetOpt(&GetState, &Val), 405, 1);
     CHECK(VALID_PTR(Val.psz) && !strcmp(Val.psz, "firstvalue"));
-    CHECK(GetState.uIndex == UINT64_MAX);
+    CHECK(GetState.uIndex == UINT32_MAX);
     CHECK_GETOPT(RTGetOptFetchValue(&GetState, &Val, RTGETOPT_REQ_STRING), VINF_SUCCESS, 1);
     CHECK(VALID_PTR(Val.psz) && !strcmp(Val.psz, "secondvalue"));
-    CHECK(GetState.uIndex == UINT64_MAX);
+    CHECK(GetState.uIndex == UINT32_MAX);
 
     CHECK_GETOPT(RTGetOpt(&GetState, &Val), 406, 2);
     CHECK(Val.u32 == 1);
@@ -333,13 +333,13 @@ int main()
 
     CHECK_GETOPT(RTGetOpt(&GetState, &Val), 407, 2);
     CHECK(Val.u32 == 1);
-    CHECK(GetState.uIndex == UINT64_MAX);
+    CHECK(GetState.uIndex == UINT32_MAX);
     CHECK_GETOPT(RTGetOptFetchValue(&GetState, &Val, RTGETOPT_REQ_UINT32), VINF_SUCCESS, 1);
     CHECK(Val.u32 == 12);
-    CHECK(GetState.uIndex == UINT64_MAX);
+    CHECK(GetState.uIndex == UINT32_MAX);
     CHECK_GETOPT(RTGetOptFetchValue(&GetState, &Val, RTGETOPT_REQ_STRING), VINF_SUCCESS, 1);
     CHECK(VALID_PTR(Val.psz) && !strcmp(Val.psz, "thirdvalue"));
-    CHECK(GetState.uIndex == UINT64_MAX);
+    CHECK(GetState.uIndex == UINT32_MAX);
 
     /* bool on/off tests */
     CHECK_GETOPT(RTGetOpt(&GetState, &Val), 408, 2);
